src/requirements.txt

fastapi[all]
python-jose[cryptography]
passlib[bcrypt]
uvicorn[all]
pydantic[all]